+++ crates/cxx-qt-gen/src/parser/method.rs
@@ -2,6 +2,7 @@
 // SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
+use crate::parser::CaseConversion;
 use crate::{
     naming::Name,
     parser::{check_safety, parameter::ParsedFunctionParameter, require_attributes},
@@ -70,7 +71,7 @@ impl ParsedMethod {
     pub fn mock_qinvokable(method: &ForeignItemFn) -> Self {
         Self {
             is_qinvokable: true,
-            ..Self::parse(method.clone(), Safety::Safe).unwrap()
+            ..Self::parse(method.clone(), Safety::Safe, CaseConversion::none()).unwrap()
         }
     }
 
@@ -101,9 +102,9 @@ impl ParsedMethod {
         Self { specifiers, ..self }
     }
 
-    pub fn parse(method: ForeignItemFn, safety: Safety) -> Result<Self> {
+    pub fn parse(method: ForeignItemFn, safety: Safety, auto_case: CaseConversion) -> Result<Self> {
         check_safety(&method, &safety)?;
-        let fields = MethodFields::parse(method)?;
+        let fields = MethodFields::parse(method, auto_case)?;
         let attrs = require_attributes(&fields.method.attrs, &Self::ALLOWED_ATTRS)?;
 
         // Determine if the method is invokable
@@ -139,14 +140,15 @@ pub struct MethodFields {
 }
 
 impl MethodFields {
-    pub fn parse(method: ForeignItemFn) -> Result<Self> {
+    pub fn parse(method: ForeignItemFn, auto_case: CaseConversion) -> Result<Self> {
         let self_receiver = foreignmod::self_type_from_foreign_fn(&method.sig)?;
         let (qobject_ident, mutability) = types::extract_qobject_ident(&self_receiver.ty)?;
         let mutable = mutability.is_some();
 
         let parameters = ParsedFunctionParameter::parse_all_ignoring_receiver(&method.sig)?;
         let safe = method.sig.unsafety.is_none();
-        let name = Name::from_ident_and_attrs(&method.sig.ident, &method.attrs, None, None)?;
+        let name =
+            Name::from_ident_and_attrs(&method.sig.ident, &method.attrs, None, None, auto_case)?;
 
         Ok(MethodFields {
             method,